+++ srcs/parsing.c
@@ -6,7 +6,7 @@
 /*   By: mthiry <mthiry@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/18 14:44:46 by ccaluwe           #+#    #+#             */
-/*   Updated: 2022/07/06 15:18:41 by mthiry           ###   ########.fr       */
+/*   Updated: 2022/07/06 15:27:36 by mthiry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
